Dont break on no module
Why: When working on a repo that has multiple branches being staged
if either branch has modules that are not in the other branch it breaks.

The workaround is to pull down a copy of the db, make a new db and
change your db settings. Seems like a waste of time to me when we
can just notify that the module does not exist.